+++ packages/client/hmi-client/src/workflow/ops/model-from-equations/mod.ts
@@ -1,6 +1,6 @@
 import { ModelFromDocumentOperation as operation } from './model-from-document-operation';
-import node from './tera-model-from-document-node.vue';
-import drilldown from './tera-model-from-document-drilldown.vue';
+import node from './tera-model-from-equations-node.vue';
+import drilldown from './tera-model-from-equations-drilldown.vue';
 
 const name = operation.name;
